requirements.txt
google-generativeai>=0.3.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-asyncio>=0.21.0  # for async testing
